download elona files to prefix folder

+++ init.cpp
@@ -96,6 +96,9 @@ void backup_config_files()
         const auto to_path = filesystem::path(from_to.second);
         if (!fs::exists(to_path))
         {
+            if (!fs::exists(from_path)) {
+                throw new std::runtime_error("Original config file " + from_path.string() + " didn't exist.");
+            }
             fs::copy_file(from_path, to_path);
         }
     }
